chess_learning_system/requirements.txt
# requirements.txt - Python dependencies for Chess Learning System

# Core game engine
pygame==2.5.2

# Web API
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Numerical operations for audio and animations
numpy==1.24.3

# Image processing for asset generation
Pillow==10.0.0

# Optional dependencies for enhanced features
# Uncomment if needed:

# For text-to-speech functionality
# pyttsx3==2.90

# For advanced analytics
# matplotlib==3.7.1
# pandas==2.0.3

# For network features (future multiplayer)
# requests==2.31.0

# For better JSON handling
# jsonschema==4.17.3

# Development dependencies
# pytest==7.4.0
# black==23.3.0
# pylint==2.17.4